feat: create article detail subcomponents
